[BPF] use the single vxlan dev only when CO-RE is available
Make no changes for legacy kernels.
