+++ felix/bpf/nat/connecttime.go
@@ -29,7 +29,6 @@ import (
 
 	"github.com/projectcalico/calico/felix/bpf"
 	"github.com/projectcalico/calico/felix/bpf/bpfdefs"
-	"github.com/projectcalico/calico/felix/bpf/bpfutils"
 	"github.com/projectcalico/calico/felix/bpf/jump"
 	"github.com/projectcalico/calico/felix/bpf/libbpf"
 	"github.com/projectcalico/calico/felix/bpf/maps"
@@ -270,7 +269,7 @@ func ProgFileName(logLevel string, ipver string) string {
 	}
 
 	btf := ""
-	if bpfutils.BTFEnabled {
+	if utils.BTFEnabled {
 		btf = "_co-re"
 	}
